2019 day 9
Once again, I had to rewrite a decent chunk of the Intcode interpreter. I hope I won't have to do this again very often.

The flaw with the previous version is that it did not make it easy to allow for relative positioning when *setting* a value. I assumed that all writes were always going to be positional, and that was a flawed assumption.